pyproject.toml
[project]
name = "polaris-asap-poses"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipython>=9.0.1",
    "loguru>=0.7.3",
    "polaris-lib>=0.11.9",
    "polars>=1.24.0",
    "rdkit>=2024.9.5",
    "typeguard>=4.4.2",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "ruff>=0.9.9",
]
